hw12: move logger interface to app package
