+++ hw12_13_14_15_calendar/internal/logger/logger.go
@@ -3,6 +3,8 @@ package logger
 import (
 	"io"
 	"log"
+
+	"github.com/petrenko-alex/otus-golang-hw/hw12_13_14_15_calendar/internal/app"
 )
 
 type Level int
@@ -13,13 +15,6 @@ const (
 	Warning
 	Error
 )
-
-type Logger interface {
-	Debug(string)
-	Info(string)
-	Warning(string)
-	Error(string)
-}
 
 type SimpleLogger struct {
 	Level Level
@@ -64,7 +59,7 @@ func (l SimpleLogger) log(prefix, msg string) {
 	l.logger.Println(msg)
 }
 
-func New(level Level, dst io.Writer) Logger {
+func New(level Level, dst io.Writer) app.Logger {
 	return SimpleLogger{
 		Level:  level,
 		logger: log.New(dst, "", 0),
